First step to refactoring 'user' to 'principal'.
a12nserver has 3 types of 'principals':

* user
* app
* group

Before this patch all three of these were also called 'users' again,
which is confusing.

This is a first step to renaming everything that applies to all three of
these to 'principals'.

Eventually, this server will get 3 namespaces for these 3 concepts:

```
/user/123
/app/123
/group/123
```

This will make everything nice and clear for devs as well.