Login/Registration redirect improvements
If you hit a URL and not logged in on a12nserver, you will get a login
screen. Before this change you would always get redirected to the root
after login.

This change makes it so you go to the URL you originally intended to go.

In addition, this also allows a `?continue` query parameter to be
specified on registration endpoints. This means we can generate links
that let users to go through the registration process and afterwards go
to a url of the developers choosing.