'Add a new privilege' action
This PR adds a form for adding a new privilege, one at a time. This new
action can be called by the user in the HAL browser, or an API client.

It's a lot easier to use this by API clients compared to the 'replace
all privileges' action that exists right now.