New login flow
Since a few versions we now have a new 'login challenge' system that is
an abstraction over all the different authentication methods and
authentication factors users use.

This is used by the OAuth2 authorization-challenge flow, but the goal
was always this system to be used by every flow that lets a user log in.

This PR introduces that new login flow, but for the moment it will be
hosted under /login/experimental until this system has full feature
parity with the normal login system.